api-server/requirements.txt
Flask==1.1.2
selenium==3.141.0
requests==2.25.0
beautifulsoup4==4.9.3
python-decouple==3.3
